refactor: refactor dbaas models
